+++ src/components/EditionDetail/Edition.test.tsx
@@ -34,15 +34,35 @@ describe("Renders edition component when given valid edition", () => {
       "/advanced-search"
     );
   });
-  test("Shows edition Title in Heading", () => {
+  test("Shows work work Title in Heading", () => {
     expect(
-      screen.getByRole("heading", { name: apiEdition.data.title })
+      screen.getByRole("heading", { name: apiEdition.data.work_title })
     ).toBeInTheDocument();
   });
   test("Shows edition Subtitle", () => {
     expect(screen.getByText(apiEdition.data.sub_title)).toBeInTheDocument();
   });
-
+  test("Shows Author name in Link", () => {
+    apiEdition.data.work_authors.forEach((author) =>
+      expect(screen.getByText(author.name)).toBeInTheDocument()
+    );
+  });
+  describe("Author links redirect to correct search query", () => {
+    test("Author without viaf", () => {
+      const authorElement = screen.getByText("Edgar, John, 1876-");
+      expect(authorElement.closest("a").href).toContain(
+        "query=author%3AEdgar%2C+John%2C+1876-"
+      );
+    });
+    test("Author with viaf", () => {
+      const authorElementWithViaf = screen.getByText(
+        "Shakespeare, William- ()"
+      );
+      expect(authorElementWithViaf.closest("a").href).toContain(
+        "query=viaf%3A96994048&display=author%3AShakespeare%2C+William-+%28%29"
+      );
+    });
+  });
   test("Three cards show up in page", () => {
     expect(
       screen.getByRole("heading", { name: "Featured Copy" })